Crate coordinate1

This documentation was generated from coordinate crate version 0.0.1. The original source code can be found on github.

Features

Handle the following Resources with ease from the central hub ...

custom field def (list)

jobs (get, insert, list, patch and update)

location (list)

schedule (get, patch and update)

team (list)

worker (list)

Everything else about the coordinate v1 API can be found at the official documentation site.

Not what you are looking for ? Find all other google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.jobs().get(...).doit()
let r = hub.jobs().update(...).doit()
let r = hub.jobs().patch(...).doit()
let r = hub.jobs().list(...).doit()
let r = hub.jobs().insert(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

Setting up your Project

To use this library, you would put the following lines into your Cargo.toml file:

[dependencies]
coordinate1 = "0.0.1"

A complete example

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "rustc-serialize" as rustc_serialize;
extern crate coordinate1;
use coordinate1::Job;
use coordinate1::cmn::Result;
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use coordinate1::Coordinate;
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and 
// `client_secret`, among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow.
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about 
// what's going on. You probably want to bring in your own `TokenStorage` to persist tokens and
// retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Coordinate::new(hyper::Client::new(), auth);
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req: Job = Default::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.jobs().update(&req, "teamId", "jobId")
             .title("nonumy")
             .progress("dolores")
             .note("gubergren")
             .lng(0.0653431304201)
             .lat(0.699208331616)
             .customer_phone_number("ea")
             .customer_name("no")
             .add_custom_field("justo")
             .assignee("justo")
             .address("et")
             .doit();
 
match result {
    Result::HttpError(err) => println!("HTTPERROR: {:?}", err),
    Result::MissingAPIKey => println!("Missing API Key"),
    Result::MissingToken => println!("Missing Token"),
    Result::Failure(_) => println!("General Failure (Response doesn't print)"),
    Result::FieldClash(clashed_field) => println!("FIELD CLASH: {:?}", clashed_field),
    Result::Success(_) => println!("Success (value doesn't print)"),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

About Customization/Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

About Parts

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifyable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

About Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are borrowed

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Modules

cmn

Macros

map!
This macro is advertised in the documentation, which is why we deliver it as well

Structs

Coordinate
Central instance to access all Coordinate related resource activities

CustomField
Custom field.

This type is not used in any activity, and only used as part of another schema.

CustomFieldDef
Custom field definition.

This type is not used in any activity, and only used as part of another schema.

CustomFieldDefListMethodBuilder
Retrieves a list of custom field definitions for a team.

CustomFieldDefListResponse
Collection of custom field definitions for a team.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list custom field def (response)

CustomFieldDefMethodsBuilder
A builder providing access to all methods supported on customFieldDef resources. It is not used directly, but through the Coordinate hub.

CustomFields
Collection of custom fields.

This type is not used in any activity, and only used as part of another schema.

EnumItemDef
Enum Item definition.

This type is not used in any activity, and only used as part of another schema.

Job
A job.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get jobs (response)

update jobs (request|response)

patch jobs (request|response)

list jobs (none)

insert jobs (request|response)

JobChange
Change to a job. For example assigning the job to a different worker.

This type is not used in any activity, and only used as part of another schema.

JobGetMethodBuilder
Retrieves a job, including all the changes made to the job.

JobInsertMethodBuilder
Inserts a new job. Only the state field of the job should be set.

JobListMethodBuilder
Retrieves jobs created or modified since the given timestamp.

JobListResponse
Response from a List Jobs request.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list jobs (response)

JobMethodsBuilder
A builder providing access to all methods supported on job resources. It is not used directly, but through the Coordinate hub.

JobPatchMethodBuilder
Updates a job. Fields that are set in the job state will be updated. This method supports patch semantics.

JobState
Current state of a job.

This type is not used in any activity, and only used as part of another schema.

JobUpdateMethodBuilder
Updates a job. Fields that are set in the job state will be updated.

Location
Location of a job.

This type is not used in any activity, and only used as part of another schema.

LocationListMethodBuilder
Retrieves a list of locations for a worker.

LocationListResponse
Response from a List Locations request.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list location (response)

LocationMethodsBuilder
A builder providing access to all methods supported on location resources. It is not used directly, but through the Coordinate hub.

LocationRecord
Recorded location of a worker.

This type is not used in any activity, and only used as part of another schema.

Schedule
Job schedule.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

update schedule (request|response)

patch schedule (request|response)

get schedule (response)

ScheduleGetMethodBuilder
Retrieves the schedule for a job.

ScheduleMethodsBuilder
A builder providing access to all methods supported on schedule resources. It is not used directly, but through the Coordinate hub.

SchedulePatchMethodBuilder
Replaces the schedule of a job with the provided schedule. This method supports patch semantics.

ScheduleUpdateMethodBuilder
Replaces the schedule of a job with the provided schedule.

Team
A Coordinate team.

This type is not used in any activity, and only used as part of another schema.

TeamListMethodBuilder
Retrieves a list of teams for a user.

TeamListResponse
Response from a List Teams request.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list team (response)

TeamMethodsBuilder
A builder providing access to all methods supported on team resources. It is not used directly, but through the Coordinate hub.

TokenPagination
Pagination information.

This type is not used in any activity, and only used as part of another schema.

Worker
A worker in a Coordinate team.

This type is not used in any activity, and only used as part of another schema.

WorkerListMethodBuilder
Retrieves a list of workers in a team.

WorkerListResponse
Response from a List Workers request.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list worker (response)

WorkerMethodsBuilder
A builder providing access to all methods supported on worker resources. It is not used directly, but through the Coordinate hub.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.