Struct coordinate1::TeamMethodsBuilder

pub struct TeamMethodsBuilder<'a, C, NC, A> where NC: 'a, C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on team resources. It is not used directly, but through the Coordinate hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "rustc-serialize" as rustc_serialize;
extern crate coordinate1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Coordinate::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *MethodBuilders*
// like `list(...)`
// to build up your call.
let rb = hub.team();

Methods

impl<'a, C, NC, A> TeamMethodsBuilder<'a, C, NC, A>

fn list(&self) -> TeamListMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Retrieves a list of teams for a user.